Fix the fork() return value explanation
You had that one wrong / reversed.

On the parent fork returns the pid of the child. Makes sense, as this pid is very often used, to wait for a child for example.

In the child, zero is returned. Makes sense, as the pid of self is seldom needed. Process can get it via the getpid() call.