Clarify description of forked process
This is my first edit to the article. It is a nitpick, so I separate
it into its own commit, but I believe there should always be a
conceptual separation between a program (the text on a page) and an
instance of the program (the process running). If nothing else, this
edit will improve the fluency of the paragraph.